Fix static files not being served
URLs beginning with /static in Flask by default are served from a
folder called "static". Since index.html makes requests to /static/css/...,
this was causing it to check for the file in a folder called "static", which
does not exist.

Explicitly setting the static_folder parameter of the app to None fixes this.